requirements.txt
python-kasa
PyQt6
qasync